chore:move function
